fix update
fixing back button and camera
